refactor: move repeated functions to global-utils
