+++ web-react-ts/src/pages/services/record-service/ServiceForm.tsx
@@ -7,7 +7,11 @@ import { useNavigate } from 'react-router'
 import { Col, Container, Row } from 'react-bootstrap'
 import { HeadingPrimary } from 'components/HeadingPrimary/HeadingPrimary'
 import SubmitButton from 'components/formik/SubmitButton'
-import { throwToSentry } from 'global-utils'
+import {
+  checkIfArrayHasRepeatingValues,
+  parseForeignCurrency,
+  throwToSentry,
+} from 'global-utils'
 import { getMondayThisWeek } from 'jd-date-utils'
 import { ChurchContext } from 'contexts/ChurchContext'
 import { Church, ChurchLevel } from 'global-types'
@@ -85,42 +89,6 @@ const ServiceForm = ({
       .min(2, 'You must have at least two treasurers')
       .of(Yup.string().required('Please pick a name from the dropdown')),
   })
-
-  const checkIfArrayHasRepeatingValues = (array: any[]) => {
-    const sortedArray = array.sort()
-    for (let i = 0; i < sortedArray.length - 1; i++) {
-      if (sortedArray[i + 1] === sortedArray[i]) {
-        return true
-      }
-    }
-    return false
-  }
-
-  const parseForeignCurrency = (string: string) => {
-    const nonOptions = [
-      0,
-      '',
-      'o',
-      '0',
-      '00',
-      '0.0',
-      '0.00',
-      'Gh',
-      'ghana cedis',
-      'No',
-      'None',
-      'n/a',
-      'N/A',
-      'NA',
-      '-',
-    ]
-
-    if (nonOptions.includes(string.toLowerCase().trim())) {
-      return null
-    }
-
-    return null
-  }
 
   const onSubmit = (
     values: FormOptions,
